Huge ass commit featuring loads of stuff
 - Add static commandaliases to all command classes
 - Move IsInsideJail & JailedPlayers list to ExentionMethods
 - New PlayerFile storage format, including new serializable classes
 - New Utility class with utilitary Methods
 - Minor optimizations

+++ AdminToolbox/AdminToolbox/Commands/Player/KillCommand.cs
@@ -6,6 +6,8 @@ using System.Collections.Generic;
 
 namespace AdminToolbox.Command
 {
+	using API;
+	using API.Extentions;
 	class KillCommand : ICommandHandler
 	{
 		private readonly AdminToolbox plugin;
@@ -14,47 +16,54 @@ namespace AdminToolbox.Command
 
 		public KillCommand(AdminToolbox plugin) => this.plugin = plugin;
 		public string GetCommandDescription() => "Kills the targeted player";
-		public string GetUsage() => "(KILL / SLAY) [PLAYER]";
+		public string GetUsage() => "(" + string.Join(" / ", CommandAliases) + ") [PLAYER]";
+
+		public static readonly string[] CommandAliases = new string[] {  "SLAY", "KILL" };
 
 		public string[] OnCall(ICommandSender sender, string[] args)
 		{
-			if (Server.GetPlayers().Count < 1)
-				return new string[] { "The server is empty!" };
-
-			Player caller = (sender is Player send) ? send : null;
-			DamageType killType = (DamageType)Config.GetIntValue("admintoolbox_slaycommand_killtype", 0, true);
-
-			if (args.Length > 0)
+			if (sender.IsPermitted(CommandAliases, out string[] deniedReply))
 			{
-				if (args[0].ToLower() == "all" || args[0].StartsWith("*"))
+				if (Server.GetPlayers().Count < 1)
+					return new string[] { "The server is empty!" };
+
+				Player caller = (sender is Player send) ? send : null;
+				DamageType killType = (DamageType)Config.GetIntValue("admintoolbox_slaycommand_killtype", 0, true);
+
+				if (args.Length > 0)
 				{
-					int playerNum = 0;
-					foreach(Player p in Server.GetPlayers().Where(pl => pl.PlayerId != (caller != null ? caller.PlayerId : -1) 
-					&& !pl.GetGodmode() &&
-					(AdminToolbox.ATPlayerDict.ContainsKey(pl.SteamId) ? !AdminToolbox.ATPlayerDict[pl.SteamId].godMode : true) 
-					&& pl.TeamRole.Team != Smod2.API.Team.SPECTATOR).ToList())
+					if (args[0].ToLower() == "all" || args[0].StartsWith("*"))
 					{
-						p.Kill(killType);
-						playerNum++;
+						int playerNum = 0;
+						foreach (Player p in Server.GetPlayers().Where(pl => pl.PlayerId != (caller != null ? caller.PlayerId : -1)
+						 && !pl.GetGodmode() &&
+						 (AdminToolbox.ATPlayerDict.ContainsKey(pl.SteamId) ? !AdminToolbox.ATPlayerDict[pl.SteamId].godMode : true)
+						 && pl.TeamRole.Team != Smod2.API.Team.SPECTATOR).ToList())
+						{
+							p.Kill(killType);
+							playerNum++;
+						}
+						if (caller != null && !string.IsNullOrEmpty(caller.Name) && caller.Name.ToLower() != "server") plugin.Info(caller.Name + " ran the \"SLAY\" command on: " + playerNum + " players");
+						return new string[] { playerNum + " players has been slain!" };
 					}
-					if (caller != null && !string.IsNullOrEmpty(caller.Name) && caller.Name.ToLower() != "server") plugin.Info(caller.Name + " ran the \"SLAY\" command on: " + playerNum + " players");
-					return new string[] { playerNum + " players has been slain!" };
-				}
-				Player myPlayer = API.GetPlayerFromString.GetPlayer(args[0]);
-				if (myPlayer == null) { return new string[] { "Couldn't get player: " + args[0] }; }
-				if (myPlayer.TeamRole.Role != Role.SPECTATOR)
-				{
-					if (caller != null && !string.IsNullOrEmpty(caller.Name) && caller.Name.ToLower() != "server") plugin.Info(caller.Name + " ran the \"SLAY\" command on: " + myPlayer.Name);
-					myPlayer.Kill(killType);
-					return new string[] { myPlayer.Name + " has been slain!" };
+					Player myPlayer = GetPlayerFromString.GetPlayer(args[0]);
+					if (myPlayer == null) { return new string[] { "Couldn't get player: " + args[0] }; }
+					if (myPlayer.TeamRole.Role != Role.SPECTATOR)
+					{
+						if (caller != null && !string.IsNullOrEmpty(caller.Name) && caller.Name.ToLower() != "server") plugin.Info(caller.Name + " ran the \"SLAY\" command on: " + myPlayer.Name);
+						myPlayer.Kill(killType);
+						return new string[] { myPlayer.Name + " has been slain!" };
+					}
+					else
+						return new string[] { myPlayer.Name + " is already dead!" };
 				}
 				else
-					return new string[] { myPlayer.Name + " is already dead!" };
+				{
+					return new string[] { GetUsage() };
+				}
 			}
 			else
-			{
-				return new string[] { GetUsage() };
-			}
+				return deniedReply;
 		}
 	}
 }
